+++ index.js
@@ -1,5 +1,5 @@
 const server = require ('./src/server')
-const dbConnect = require ('./src/server')
+const dbConnect = require ('./src/lib/db')
 
 dbConnect()
     .then(() => {
