+++ src/server.js
@@ -1,10 +1,10 @@
 const express = require('express')
-const writersRouter = require('./routers/writers')
+// const writersRouter = require('./routers/writers')
 
 const server = express()
 
 server.use(express.json())
 
-server.use('/writers', writersRouter)
+// server.use('/writers', writersRouter)
 
 module.exports = server
